Initial psxgte cleanup
